lmdb: move to seperate source file and fix configure
